refactor: consolidate 40 test binaries into single integration binary
Each .rs file in tests/ was compiled as a separate binary linking the
full dependency tree (aws-sdk-s3, tokio, hyper, rustls, etc.), causing
linker OOM on CI runners. Consolidate into a single tests/integration.rs
entry point with module declarations.

Also reverts the CARGO_BUILD_JOBS=1 and swap space CI workarounds since
they're no longer needed.
