+++ .github/workflows/ci.yml
@@ -37,16 +37,8 @@ jobs:
   test:
     name: test
     runs-on: ubuntu-latest
-    env:
-      CARGO_BUILD_JOBS: 1
     steps:
       - uses: actions/checkout@v4
-      - name: Add swap space
-        run: |
-          sudo fallocate -l 8G /swapfile
-          sudo chmod 600 /swapfile
-          sudo mkswap /swapfile
-          sudo swapon /swapfile
       - uses: actions/cache@v4
         with:
           path: |
