+++ crates/loon-testkit/tests/integration/client_observe_delete_move.rs
