+++ crates/loon-testkit/tests/integration/local_only_binding.rs
